feat: change populate script to be conditional on if file exists

+++ populate.sh
@@ -1,4 +1,8 @@
 #!/bin/sh
 
 mkdir -p ~/.config/sway
-ln ./sway.conf ~/.config/sway/config
+if [ ! -e ~/.config/sway/config ]
+then
+	echo "Creating symlink for sway config"
+	ln ./sway.conf ~/.config/sway/config
+fi
